refactor(pdf): rename pdf related directories and files
